+++ packages/contracts/deployments/engine/V3/studio/mainnet/2025-04-03-deployment-config.ts
@@ -1,0 +1,160 @@
+// This file is used to configure the deployment of Engine and Engine Flex contracts
+// It is intended to be imported by the batch engine factory deployer by running `deploy:v3-engine:mainnet`, `deploy:v3-engine:staging` or `deploy:v3-engine:dev`.
+
+export const deployNetworkConfiguration = {
+  network: "mainnet",
+  // environment is only used for metadata purposes, and is not used in the deployment process
+  // Please set to "dev", "staging", or "mainnet", arbitrum as appropriate
+  environment: "mainnet",
+  useLedgerSigner: true,
+  // ONLY Gnosis Safe is supported for Studio deployments at this time
+  useGnosisSafe: true,
+  safeAddress: "0x52119BB73Ac8bdbE59aF0EEdFd4E4Ee6887Ed2EA",
+  transactionServiceUrl: "https://safe-transaction-mainnet.safe.global",
+  // set the transaction hash after successful execution of the batch creation
+  transactionHash:
+    "0x149db9a28cde0d5e40f420487609aadd5f359cf8aa77e022a9caf8442649be1b",
+};
+
+export const deployConfigDetailsArray = [
+  {
+    // productClass is either "Engine" or "Studio", and used to validate render provider payment address
+    productClass: "Studio",
+    // 0 for Engine, 1 for Engine Flex
+    engineCoreContractType: 0,
+    // prod-only: see efficient_addresses...txt in network directories for a list of efficient salts
+    salt: "0x0000000000000000000000000000000000000000f7634f1ab331c3a4cd0500e0",
+    // INCREMENT THESE NUMBERS FOR EACH NEW STUDIO DEPLOYMENT
+    tokenName: "Art Blocks Studio | 74",
+    tokenTicker: "ABSTUDIO_74",
+    // optionally define this for improved readability and searchability in the deployment logs
+    artistName: "Emily Nicoll, 747",
+    // set to the address you want to transfer the superAdmin role to
+    // (can be the zero address if you have an existing Admin ACL Contract and are not transferring)
+    newSuperAdminAddress: "0x130Fa4Ac8EdeD3D7667A4Ba550967b5f7fEcf32C",
+    // see /scripts/util/constants.ts::MAIN_CONFIG for the correct address if prod deployment
+    renderProviderAddress: "0x036F3D03C1ccdde1878F01607922EA12110Ee9Bd",
+    // platform provider address should be set to the zero address for Studio deployments
+    platformProviderAddress: "0x0000000000000000000000000000000000000000",
+    startingProjectId: 0,
+    // autoApproveArtistSplitProposals should be true for Studio deployments
+    autoApproveArtistSplitProposals: true,
+    // nullPlatformProvider must be true for Studio deployments
+    nullPlatformProvider: true,
+    // allowArtistProjectActivation must be true for Studio deployments
+    allowArtistProjectActivation: true,
+    // if you want to use an existing admin ACL, set the address here (otherwise set as the zero address 0x0000000000000000000000000000000000000000 to deploy a new one)
+    adminACLContract: "0x0000000000000000000000000000000000000000",
+    // optionally define this to set default vertical name for the contract after deployment.
+    // if not defined, the default vertical name will be "unassigned".
+    // common values include `studio`, `fullyonchain`, `flex`, or partnerships like `artblocksxpace`.
+    // also note that if you desire to create a new vertical, you will need to add the vertical name to the
+    // `project_verticals` table in the database before running this deploy script.
+    defaultVerticalName: "studio",
+  },
+  {
+    // productClass is either "Engine" or "Studio", and used to validate render provider payment address
+    productClass: "Studio",
+    // 0 for Engine, 1 for Engine Flex
+    engineCoreContractType: 0,
+    // prod-only: see efficient_addresses...txt in network directories for a list of efficient salts
+    salt: "0x0000000000000000000000000000000000000000f7634f1ab331228d850500c0",
+    // INCREMENT THESE NUMBERS FOR EACH NEW STUDIO DEPLOYMENT
+    tokenName: "Art Blocks Studio | 75",
+    tokenTicker: "ABSTUDIO_75",
+    // optionally define this for improved readability and searchability in the deployment logs
+    artistName: "Shaunaq",
+    // set to the address you want to transfer the superAdmin role to
+    // (can be the zero address if you have an existing Admin ACL Contract and are not transferring)
+    newSuperAdminAddress: "0x57E3865513815767Bc1bad2Ac69a5B899D4aC459",
+    // see /scripts/util/constants.ts::MAIN_CONFIG for the correct address if prod deployment
+    renderProviderAddress: "0x036F3D03C1ccdde1878F01607922EA12110Ee9Bd",
+    // platform provider address should be set to the zero address for Studio deployments
+    platformProviderAddress: "0x0000000000000000000000000000000000000000",
+    startingProjectId: 0,
+    // autoApproveArtistSplitProposals should be true for Studio deployments
+    autoApproveArtistSplitProposals: true,
+    // nullPlatformProvider must be true for Studio deployments
+    nullPlatformProvider: true,
+    // allowArtistProjectActivation must be true for Studio deployments
+    allowArtistProjectActivation: true,
+    // if you want to use an existing admin ACL, set the address here (otherwise set as the zero address 0x0000000000000000000000000000000000000000 to deploy a new one)
+    adminACLContract: "0x0000000000000000000000000000000000000000",
+    // optionally define this to set default vertical name for the contract after deployment.
+    // if not defined, the default vertical name will be "unassigned".
+    // common values include `studio`, `fullyonchain`, `flex`, or partnerships like `artblocksxpace`.
+    // also note that if you desire to create a new vertical, you will need to add the vertical name to the
+    // `project_verticals` table in the database before running this deploy script.
+    defaultVerticalName: "studio",
+  },
+  {
+    // productClass is either "Engine" or "Studio", and used to validate render provider payment address
+    productClass: "Studio",
+    // 0 for Engine, 1 for Engine Flex
+    engineCoreContractType: 1,
+    // prod-only: see efficient_addresses...txt in network directories for a list of efficient salts
+    salt: "0x0000000000000000000000000000000000000000a73bf3db9deb2769150c0010",
+    // INCREMENT THESE NUMBERS FOR EACH NEW STUDIO DEPLOYMENT
+    tokenName: "Art Blocks Studio | 76",
+    tokenTicker: "ABSTUDIO_76",
+    // optionally define this for improved readability and searchability in the deployment logs
+    artistName: "basement",
+    // set to the address you want to transfer the superAdmin role to
+    // (can be the zero address if you have an existing Admin ACL Contract and are not transferring)
+    newSuperAdminAddress: "0x2574c77a694700b7baB562fefeB9Ce93DB5A097A",
+    // see /scripts/util/constants.ts::MAIN_CONFIG for the correct address if prod deployment
+    renderProviderAddress: "0x036F3D03C1ccdde1878F01607922EA12110Ee9Bd",
+    // platform provider address should be set to the zero address for Studio deployments
+    platformProviderAddress: "0x0000000000000000000000000000000000000000",
+    startingProjectId: 0,
+    // autoApproveArtistSplitProposals should be true for Studio deployments
+    autoApproveArtistSplitProposals: true,
+    // nullPlatformProvider must be true for Studio deployments
+    nullPlatformProvider: true,
+    // allowArtistProjectActivation must be true for Studio deployments
+    allowArtistProjectActivation: true,
+    // if you want to use an existing admin ACL, set the address here (otherwise set as the zero address 0x0000000000000000000000000000000000000000 to deploy a new one)
+    adminACLContract: "0x0000000000000000000000000000000000000000",
+    // optionally define this to set default vertical name for the contract after deployment.
+    // if not defined, the default vertical name will be "unassigned".
+    // common values include `studio`, `fullyonchain`, `flex`, or partnerships like `artblocksxpace`.
+    // also note that if you desire to create a new vertical, you will need to add the vertical name to the
+    // `project_verticals` table in the database before running this deploy script.
+    defaultVerticalName: "studio",
+  },
+  {
+    // productClass is either "Engine" or "Studio", and used to validate render provider payment address
+    productClass: "Studio",
+    // 0 for Engine, 1 for Engine Flex
+    engineCoreContractType: 0,
+    // prod-only: see efficient_addresses...txt in network directories for a list of efficient salts
+    salt: "0x0000000000000000000000000000000000000000f7634f1ab331aef0e2040000",
+    // INCREMENT THESE NUMBERS FOR EACH NEW STUDIO DEPLOYMENT
+    tokenName: "Art Blocks Studio | 77",
+    tokenTicker: "ABSTUDIO_77",
+    // optionally define this for improved readability and searchability in the deployment logs
+    artistName: "Rafaël Rozendaal",
+    // set to the address you want to transfer the superAdmin role to
+    // (can be the zero address if you have an existing Admin ACL Contract and are not transferring)
+    newSuperAdminAddress: "0x0F441cFaD93287109F5eF834bF52F4aaaa8d8ffa",
+    // see /scripts/util/constants.ts::MAIN_CONFIG for the correct address if prod deployment
+    renderProviderAddress: "0x036F3D03C1ccdde1878F01607922EA12110Ee9Bd",
+    // platform provider address should be set to the zero address for Studio deployments
+    platformProviderAddress: "0x0000000000000000000000000000000000000000",
+    startingProjectId: 0,
+    // autoApproveArtistSplitProposals should be true for Studio deployments
+    autoApproveArtistSplitProposals: true,
+    // nullPlatformProvider must be true for Studio deployments
+    nullPlatformProvider: true,
+    // allowArtistProjectActivation must be true for Studio deployments
+    allowArtistProjectActivation: true,
+    // if you want to use an existing admin ACL, set the address here (otherwise set as the zero address 0x0000000000000000000000000000000000000000 to deploy a new one)
+    adminACLContract: "0x0000000000000000000000000000000000000000",
+    // optionally define this to set default vertical name for the contract after deployment.
+    // if not defined, the default vertical name will be "unassigned".
+    // common values include `studio`, `fullyonchain`, `flex`, or partnerships like `artblocksxpace`.
+    // also note that if you desire to create a new vertical, you will need to add the vertical name to the
+    // `project_verticals` table in the database before running this deploy script.
+    defaultVerticalName: "studio",
+  },
+];