+++ packages/contracts/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @artblocks/contracts
 
+## 1.3.2
+
+### Patch Changes
+
+- 713e914: Updated PMPV0 with improved provenance event data
+
+## 1.3.1
+
+### Patch Changes
+
+- e03afaf: gas efficiency updates to PMPV0
+
+## 1.3.0
+
+### Minor Changes
+
+- bc4019a: post mint parameters and associated contracts
+
 ## 1.2.1
 
 ### Patch Changes